+++ lib/gds_api/exceptions.rb
@@ -1,17 +1,6 @@
 module GdsApi
   # Abstract error class
-  class BaseError < StandardError
-    # Give Sentry extra context about this event
-    # https://docs.sentry.io/clients/ruby/context/
-    def raven_context
-      {
-        # Make Sentry group exceptions by type instead of message, so all
-        # exceptions like `GdsApi::TimedOutException` will get grouped as one
-        # error and not an error per URL.
-        fingerprint: [self.class.name],
-      }
-    end
-  end
+  class BaseError < StandardError; end
 
   class EndpointNotFound < BaseError; end
   class TimedOutException < BaseError; end